New game level version spec
- all positions are pixel precision (i32)
- keep previous level version in level::v0

+++ src/level/info.rs
@@ -2,28 +2,28 @@ use na::Vector2;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct GemInfo {
-    pub pos: Vector2<f32>,
+    pub pos: Vector2<i32>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct PumpInfo {
-    pub pos: Vector2<f32>,
+    pub pos: Vector2<i32>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct MineInfo {
-    pub pos: Vector2<f32>,
+    pub pos: Vector2<i32>,
 }
 
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct WallInfo {
-    pub pos: Vector2<f32>,
-    pub dim: Vector2<f32>,
+    pub pos: Vector2<i32>,
+    pub dim: Vector2<i32>,
     #[serde(default)] pub texture_id: u32,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct FinishInfo {
-    pub pos: Vector2<f32>,
+    pub pos: Vector2<i32>,
     #[serde(default)] pub gems_required: u32,
 }
